BOILERPLATE/BOILERPLATE_SQL/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9  # PostgreSQL driver
alembic>=1.13.1  # Database migrations

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Validation and settings
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP client
httpx>=0.26.0

# Development
python-dotenv>=1.0.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# Optional: MySQL support
# pymysql>=1.1.0

# Optional: Async PostgreSQL
# asyncpg>=0.29.0
